+++ CHANGELOG.md
@@ -144,6 +144,9 @@ This is more compliant with RFC 6066, and therefore more compatible with stricte
       - Usage of the wildcard '*' 'Contact' header field was not correctly handled (considered as a bad request).
       - Requests of type 'REGISTER' were rejected if they did not contain an expiry value (it is now compliant with
         RFC3261).
+    - **NatHelper:**
+      - The `contact-verified-param` parameter has been renamed but not deprecated, causing problems at startup if the
+        parameter was defined in the configuration file.
 - **B2bua/SIP Bridge**: authentication information for deleted accounts was removed too early, preventing accounts from
   properly unregistering to the registrar.
 - **EventLogs (`flexiapi` logger only):** Two events (INVITE, MESSAGE, etc.) sharing the same Call-ID, the same user